Move FlashROM out of GPIO stuff because it isn't
