+++ src/gba/overrides.c
@@ -375,11 +375,9 @@ void GBAOverrideApply(struct GBA* gba, const struct GBACartridgeOverride* overri
 		} else {
 			gba->memory.hw.devices &= ~HW_GB_PLAYER_DETECTION;
 		}
-		
-		if (override->hardware & HW_FLASHROM) {
-			gba->memory.hw.devices |= HW_FLASHROM;
-		}
 	}
+	
+	gba->memory.flashrom.type = override->flashromtype;
 
 	if (override->idleLoop != IDLE_LOOP_NONE) {
 		gba->idleLoop = override->idleLoop;
